+++ common/ip-packet-requests/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "nym-ip-packet-requests"
+version = "0.1.0"
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+documentation.workspace = true
+edition.workspace = true
+license.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bincode = "1.3.3"
+bytes = "1.5.0"
+nym-service-providers-common = { path = "../../service-providers/common" }
+nym-sphinx = { path = "../nymsphinx" }
+serde = { workspace = true, features = ["derive"] }
