+++ common/tun/src/tun_task_channel.rs
@@ -1,6 +1,5 @@
 use std::time::Duration;
 
-use nym_service_providers_common::interface;
 use tokio::sync::mpsc::{
     self,
     error::{SendError, SendTimeoutError, TrySendError},
@@ -80,36 +79,4 @@ pub(crate) fn tun_task_response_channel() -> (TunTaskResponseTx, TunTaskResponse
         TunTaskResponseTx(tun_task_tx),
         TunTaskResponseRx(tun_task_rx),
     )
-}
-
-pub type IpPacketRouterRequest = interface::Request<TaggedIpPacket>;
-pub type IpPacketRouterResponse = interface::Response<IpPacket>;
-
-#[derive(serde::Serialize, serde::Deserialize)]
-pub struct TaggedIpPacket {
-    pub packet: bytes::Bytes,
-    pub return_address: nym_sphinx::addressing::clients::Recipient,
-    pub return_mix_hops: Option<u8>,
-    pub return_mix_delays: Option<f64>,
-}
-
-#[derive(serde::Serialize, serde::Deserialize)]
-pub struct IpPacket {
-    pub packet: bytes::Bytes,
-}
-
-impl TaggedIpPacket {
-    pub fn from_message(
-        message: &nym_sphinx::receiver::ReconstructedMessage,
-    ) -> Result<Self, bincode::Error> {
-        use bincode::Options;
-        make_bincode_serializer().deserialize(&message.message)
-    }
-}
-
-fn make_bincode_serializer() -> impl bincode::Options {
-    use bincode::Options;
-    bincode::DefaultOptions::new()
-        .with_big_endian()
-        .with_varint_encoding()
 }
